Разделил StatisticsRepository на два репозитория
Теперь есть 2 отдельных репозитория ChangeStatisticsRepository, используемый в тренажерах, и GetStatisticsRepository, используемый на экране профиля для отображения статистики
