+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/trainer/ChangeStatisticsRepository.kt
@@ -1,0 +1,7 @@
+package com.designdrivendevelopment.kotelok.trainer
+
+interface ChangeStatisticsRepository {
+    suspend fun addSuccessfulResultToWordDef(wordDefinitionsId: Long)
+
+    suspend fun addFailedResultToWordDef(wordDefinitionsId: Long)
+}
